requirements.txt
alembic==1.4.3
aniso8601==8.0.0
astroid==2.4.2
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
mccabe==0.6.1
psycopg2==2.8.6
pylint==2.6.0
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
